requirements.txt
pytest
requests
requests-auth
allure-pytest